+++ src/candlewick/core/Components.cpp
@@ -2,17 +2,19 @@
 #include <algorithm>
 
 namespace candlewick {
+bool MeshMaterialComponent::hasTransparency() const {
+  return std::ranges::any_of(
+      materials, [](auto &material) { return material.baseColor.w() < 1.0f; });
+}
+
 bool updateTransparencyClassification(entt::registry &reg, entt::entity entity,
                                       const MeshMaterialComponent &mmc) {
-  bool hasTransparency = std::ranges::any_of(mmc.materials, [](auto &material) {
-    return material.baseColor.w() < 1.0f;
-  });
-
-  if (hasTransparency) {
+  bool is_transparent = mmc.hasTransparency();
+  if (is_transparent) {
     reg.remove<Opaque>(entity);
   } else {
     reg.emplace_or_replace<Opaque>(entity);
   }
-  return hasTransparency;
+  return is_transparent;
 }
 } // namespace candlewick